adding nonces and validating input on exam & grades pages
git-svn-id: https://plugins.svn.wordpress.org/watu/trunk@1290987 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ controllers/grades.php
@@ -6,21 +6,29 @@ function watu_grades() {
    $quiz = $wpdb->get_row($wpdb->prepare("SELECT ID, name FROM ".WATU_EXAMS." WHERE ID=%d", $_GET['quiz_id']));
    if(empty($quiz->ID)) wp_die(__('Unrecognized quiz ID', 'watu'));
    
-   if(!empty($_POST['add'])) {
+   if(!empty($_POST['add']) or !empty($_POST['save'])) {
+   	$gtitle = sanitize_text_field($_POST['gtitle']);	   
+	   $gfrom = intval($_POST['gfrom']);
+	   $gto = intval($_POST['gto']);
+   }
+   
+   if(!empty($_POST['add']) and check_admin_referer('watu_grades')) {
+   	$gdesc = $_POST['gdesc'];
    	$wpdb->query($wpdb->prepare("INSERT INTO ".WATU_GRADES." SET
    		exam_id=%d, gtitle=%s, gdescription=%s, gfrom=%d, gto=%d",
-   		$quiz->ID, $_POST['gtitle'], $_POST['gdesc'], $_POST['gfrom'], $_POST['gto']));
+   		$quiz->ID, $gtitle, $gdesc, $gfrom, $gto));
    	watu_redirect("admin.php?page=watu_grades&quiz_id=".$quiz->ID);	
    }
    
-   if(!empty($_POST['save'])) {
+   if(!empty($_POST['save']) and check_admin_referer('watu_grades')) {
+   	$gdesc = $_POST['gdesc'.$_POST['id']];
    	$wpdb->query($wpdb->prepare("UPDATE ".WATU_GRADES." SET
    		gtitle=%s, gdescription=%s, gfrom=%d, gto=%d WHERE ID=%d",
-   		$_POST['gtitle'], $_POST['gdesc'.$_POST['id']], $_POST['gfrom'], $_POST['gto'], $_POST['id']));
+   		$gtitle, $gdesc, $gfrom, $gto, $_POST['id']));
    	watu_redirect("admin.php?page=watu_grades&quiz_id=".$quiz->ID);
    }
    
-   if(!empty($_POST['del'])) {
+   if(!empty($_POST['del']) and check_admin_referer('watu_grades')) {
    	$wpdb->query($wpdb->prepare("DELETE FROM ".WATU_GRADES." WHERE ID=%d", $_POST['id']));
    }
